initial tox configuration
add tox configuration for testing with multiple versions of Python
and Django. Initial commit handles the following versions:

* Python 2.7, Django 1.8
* Python 2.7, Django 1.9

Python 2.6 and earlier are not supported by Django, and this
codebase does not support Python 3.x at this time.

To run tests:

    pip install tox
    tox

Note that one of the templatetag tests currently fails under tox,
due to a discrepancy in the ordering of hash attributes in
the template output.