using consisting sorting of dicts in BaseModel.to_json
As a way to simplify testing, always sort dicts during the final
serialization in the to_json method. This ensures consistent
HTML output in the rendered template, which makes it trivial
to compare to expected output.

NOTE: this change should be reviewed since it has an impact,
albeit minor, on the behavior of this method.