remove pytest as a test dependency
pytest and pytest-django invoke some magic __import__ behavior
that seems to complicate things when running tests via tox.

Since the test suite doesn't rely on any of this magic yet, this
commit removes pytest, instead relying on standard Django test
behavior.

Note that a 'runtests.py' script is added to handle setting up the
Django environment. In the future, it might be tidier to move this
code into setup.py.